requirements.txt
penman==0.6.2
networkx
sacrebleu
num2words
word2number
overrides
torch==1.3.1
h5py
boto3
pytorch-pretrained-bert==0.6.2
spacy
ftfy
nltk
conllu
pyyaml
pycorenlp
editdistance
bs4
lxml
subword-nmt
setproctitle
filelock
ijson
tokenizers
tensorflow==1.14.0
transformers
tensorboardX
ijson
filelock
scipy
setproctitle
amrlib

## requirements to extract affective features
gensim==3.8.0
joblib==0.14.1
Keras==2.2.4
Keras-Preprocessing==1.1.0
keras-self-attention==0.35.0
numpy==1.21.6
pandas==0.24.2
nltk==3.4.5
scikit-learn
tensorflow-gpu
tqdm==4.32.1
hyperopt
